changed unsigned int to unsigned short int

+++ cpp_project/character.cpp
@@ -1,6 +1,6 @@
 #include "character.h"
 
-Character::Character(Coords *hitBox, Coords *position, unsigned int speed, unsigned int maxHealth, unsigned int currentHealth) :
+Character::Character(Coords *hitBox, Coords *position, unsigned short int speed, unsigned short int maxHealth, unsigned short int currentHealth) :
     hitBox(hitBox), position(position), speed(speed), maxHealth(maxHealth), currentHealth(currentHealth)
 {
     moveAllowance = 0;
