+++ cpp_project/vampire.h
@@ -6,12 +6,12 @@
 class Vampire : public Enemy
 {
 public:
-    Vampire(Coords* hitBox, Coords* position, unsigned int speed, unsigned int maxHealth, unsigned int currentHealth, unsigned int damage, Coords* target);
+    Vampire(Coords* hitBox, Coords* position, unsigned short int speed, unsigned short int maxHealth, unsigned short int currentHealth, unsigned short int damage, Coords* target);
 
     void move(float timeDeltaMs) override;
     void moveOnlyX() override;
     void moveOnlyY() override;
-    void modifyHealth(int amount) override;
+    void modifyHealth(short int amount) override;
 
 private:
     unsigned int damage;
